feat(apis): update configuration, modifieds relation one to one  payments / bookings  and added feature save spaces

+++ src/main/java/com/eam/surfspace/persistence/repository/BookingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import com.eam.surfspace.persistence.entity.BookingEntity;
 
 @Repository
-public interface BookingRepository extends JpaRepository<BookingEntity, Long> {
+public interface BookingRepository extends JpaRepository<BookingEntity, Integer> {
     // interfaz que extiende de JpaRepository para operaciones CRUD en BookingEntity
 
 }
